mevbot/strategies/optimized_jit.py
"""Optimized JIT Liquidity Strategy"""
import numpy as np
from typing import Dict, List, Tuple, Optional
import asyncio
import logging
from web3 import Web3

from ..core.strategy_base import BaseStrategy
from ..core.memory import MemoryManager
from ..core.flash_loan import FlashLoanManager
from ..core.performance import PerformanceMonitor

logger = logging.getLogger(__name__)

class OptimizedJITStrategy(BaseStrategy):
    """JIT Liquidity Strategy with SIMD and event loop optimizations"""
    
    def __init__(
        self,
        web3: Web3,
        flash_loan_manager: FlashLoanManager,
        memory_manager: MemoryManager,
        config: Dict
    ):
        super().__init__(config)
        self.web3 = web3
        self.flash_loan_manager = flash_loan_manager
        self.memory_manager = memory_manager
        self.performance = PerformanceMonitor(config)
        
        # Pre-allocate numpy arrays for performance
        self.pool_states = np.zeros((100, 4))  # [reserve0, reserve1, price, liquidity]
        self.pending_txs = np.zeros((1000, 3))  # [value, gas_price, gas_limit]
    
    async def monitor_pools_vectorized(
        self,
        pool_addresses: List[str]
    ) -> np.ndarray:
        """Monitor multiple pools using vectorized operations"""
        async with self.performance.measure_latency('pool_updates'):
            try:
                # Batch RPC calls for efficiency
                async with self.performance.measure_latency('rpc_calls'):
                    calls = [
                        self.web3.eth.contract(
                            address=addr,
                            abi=self.config['pool_abi']
                        ).functions.getReserves().call()
                        for addr in pool_addresses
                    ]
                    
                    # Execute calls concurrently
                    results = await asyncio.gather(*calls, return_exceptions=True)
                
                async with self.performance.measure_latency('calculations'):
                    # Convert to numpy arrays for SIMD
                    reserves = np.array([
                        [r[0], r[1]] for r in results if not isinstance(r, Exception)
                    ])
                    
                    # Update success rate
                    success_count = sum(1 for r in results if not isinstance(r, Exception))
                    self.performance.update_success_rate(
                        'opportunities',
                        success_count / len(results) if results else 0
                    )
                    
                    if len(reserves) == 0:
                        return np.array([])
                    
                    # Vectorized calculations
                    prices = np.divide(reserves[:, 0], reserves[:, 1], where=reserves[:, 1]!=0)
                    liquidity = np.sqrt(np.multiply(reserves[:, 0], reserves[:, 1]))
                    
                    # Update pool states
                    self.pool_states[:len(reserves), 0] = reserves[:, 0]
                    self.pool_states[:len(reserves), 1] = reserves[:, 1]
                    self.pool_states[:len(reserves), 2] = prices
                    self.pool_states[:len(reserves), 3] = liquidity
                    
                    return self.pool_states[:len(reserves)]
                    
            except Exception as e:
                logger.error(f"Error monitoring pools: {e}")
                return np.array([])
    
    async def calculate_optimal_position(
        self,
        pool_state: np.ndarray,
        pending_tx: Dict
    ) -> Tuple[float, float]:
        """Calculate optimal position using SIMD"""
        try:
            # Extract parameters
            amount_in = float(pending_tx['value'])
            current_price = pool_state[2]
            
            # Vectorized calculations for different position sizes
            position_sizes = np.linspace(0, amount_in, 100)
            price_impacts = np.divide(position_sizes, pool_state[3])
            expected_prices = current_price * (1 - price_impacts)
            
            # Calculate expected profits
            fees = position_sizes * self.config['fee_rate']
            profits = np.multiply(position_sizes, expected_prices) - fees
            
            # Find optimal position
            optimal_idx = np.argmax(profits)
            return float(position_sizes[optimal_idx]), float(profits[optimal_idx])
            
        except Exception as e:
            logger.error(f"Error calculating position: {e}")
            return 0.0, 0.0
            
    async def execute_opportunity(self, params: Dict) -> bool:
        """Execute JIT opportunity with optimized timing"""
        try:
            async with self.performance.measure_latency('tx_submissions'):
                # Get flash loan if needed
                if params.get('needs_loan'):
                    loan_result = await self.flash_loan_manager.execute_flash_loan({
                        'token': params['token'],
                        'amount': params['amount'],
                        'target': params['pool']
                    })
                    self.performance.update_success_rate('flash_loans', bool(loan_result))
                    if not loan_result:
                        return False
                
                # Calculate gas cost
                gas_price = self.web3.eth.gas_price
                gas_limit = self.optimize_gas_vectorized(
                    gas_prices=params.get('gas_prices', []),
                    success_rates=params.get('success_rates', [])
                )
                gas_cost = float(gas_price) * float(gas_limit)
                
                # Prepare transaction
                tx = {
                    'to': params['pool'],
                    'value': params['amount'],
                    'gas': gas_limit,
                    'gasPrice': gas_price,
                    'data': params['data'],
                    'nonce': 0,  # Use fixed nonce for testing
                    'chainId': self.web3.eth.chain_id
                }
                
                logger.debug(f"Prepared transaction: {tx}")
                
                # Execute with precise timing
                async with self.memory_manager.lock:
                    try:
                        # Sign transaction
                        signed_tx = self.web3.eth.account.sign_transaction(
                            tx,
                            private_key=self.config['private_key']
                        )
                        logger.debug(f"Signed transaction: {signed_tx.rawTransaction.hex()}")
                        
                        # Send transaction
                        tx_hash = await self.web3.eth.send_raw_transaction(
                            signed_tx.rawTransaction
                        )
                        
                        success = bool(tx_hash and len(tx_hash) >= 32)
                        logger.debug(f"Transaction hash: {tx_hash.hex() if success else None}")
                        
                    except Exception as e:
                        logger.error(f"Transaction error: {e}")
                        success = False
                
                self.performance.update_success_rate('executions', success)
                
                if success:
                    # Calculate net profit (expected profit - gas cost)
                    net_profit = float(params['expected_profit']) - gas_cost / 1e18  # Convert gas cost to ETH
                    self.performance.record_profit(net_profit, gas_cost)
                    logger.info(f"Recorded profit: {net_profit} ETH (gas cost: {gas_cost/1e18} ETH)")
                
                return success
                
        except Exception as e:
            logger.error(f"Error executing opportunity: {e}")
            return False
    
    def optimize_gas_vectorized(self, gas_prices: List[int], success_rates: List[float]) -> int:
        """Optimize gas limit using historical data with vectorized operations"""
        try:
            # Debug input types
            logger.debug(f"Gas prices type: {type(gas_prices)}, Success rates type: {type(success_rates)}")
            logger.debug(f"Gas prices: {gas_prices}")
            logger.debug(f"Success rates: {success_rates}")
            
            # Convert inputs to numpy arrays if needed
            if not isinstance(gas_prices, np.ndarray):
                gas_prices = np.array(gas_prices, dtype=float)
            if not isinstance(success_rates, np.ndarray):
                success_rates = np.array(success_rates, dtype=float)
            
            # Debug array shapes
            logger.debug(f"Gas prices shape: {gas_prices.shape}, Success rates shape: {success_rates.shape}")
            
            # Calculate weighted gas limit based on success rates
            if len(gas_prices) > 0 and len(success_rates) > 0:
                weights = success_rates / np.sum(success_rates)
                logger.debug(f"Weights: {weights}")
                
                optimal_gas = int(np.sum(gas_prices * weights))
                logger.debug(f"Optimal gas before safety: {optimal_gas}")
                
                # Add safety margin (10%)
                optimal_gas = int(optimal_gas * 1.1)
                logger.debug(f"Optimal gas after safety: {optimal_gas}")
                
                # Ensure gas is within reasonable bounds
                min_gas = 21000  # Minimum for ETH transfer
                max_gas = 500000  # Maximum reasonable limit
                optimal_gas = max(min_gas, min(optimal_gas, max_gas))
                logger.debug(f"Final gas limit: {optimal_gas}")
                
                return optimal_gas
            else:
                logger.warning("Empty gas prices or success rates, using default")
                return 200000  # Default gas limit if no historical data
                
        except Exception as e:
            logger.error(f"Error optimizing gas: {e}", exc_info=True)
            return 200000  # Default gas limit on error
    
    async def _run_iteration(self):
        """Single iteration of JIT strategy"""
        try:
            # Start performance monitoring
            asyncio.create_task(self.performance.monitor_loop())
            
            # Monitor pools
            pool_states = await self.monitor_pools_vectorized(
                self.config['pool_addresses']
            )
            if len(pool_states) == 0:
                return
            
            async with self.performance.measure_latency('calculations'):
                # Analyze opportunities
                market_data = {
                    'prices': pool_states[:, 2],
                    'volumes': pool_states[:, 3],
                    'fees': np.full_like(pool_states[:, 2], self.config['fee_rate'])
                }
                
                has_opportunity, profit = self.analyze_opportunities_vectorized(
                    market_data
                )
                
                if has_opportunity and profit > self.config['min_profit']:
                    # Calculate optimal position
                    amount, expected_profit = await self.calculate_optimal_position(
                        pool_states[0],  # Use best pool
                        {'value': profit * 0.95}  # 95% of identified profit
                    )
                    
                    if expected_profit > self.config['min_profit']:
                        # Execute the opportunity
                        await self.execute_opportunity({
                            'pool': self.config['pool_addresses'][0],
                            'amount': amount,
                            'needs_loan': amount > self.config['max_own_liquidity'],
                            'gas_prices': self.config['gas_price_history'],
                            'success_rates': self.config['success_rate_history'],
                            'token': self.config['token_address'],
                            'data': self.config['interaction_data'],
                            'expected_profit': expected_profit
                        })
                    
        except Exception as e:
            logger.error(f"Error in JIT iteration: {e}")
            await asyncio.sleep(1)
